fixed window fullscreen floating. cleanup
